test for next_idx change
this adds a test for next_idx on append entry, which changes a number of tests which checked for this

in order to do this, I had to modify the existing mock callback to return 0 instead of 1.  this needs to be verified

more concerning is that the term is now changed.  the new value seems right to me, but needs to be verified